test: reorganize mobile test responsibilities
