+++ .github/workflows/web-quality.yml
@@ -68,7 +68,7 @@ jobs:
         run: pnpm build
 
   e2e:
-    name: Mobile recent home E2E
+    name: Mobile WebView E2E
     runs-on: ubuntu-latest
     timeout-minutes: 15
